app/requirements.txt
Flask==3.1.0
pymongo==4.10.1
mysql-connector-python==9.1.0